Extract Magento test stubs to dedicated files to pass Marketplace phpcs
The Magento Marketplace v5.0.0 technical validation report rejected
the package with 42 phpcs severity-10 errors, all concentrated in 5
test files. Two error families, one root cause: each test file used
bracketed namespace { ... } syntax to inline-define multiple Magento
stub classes alongside the real test class, which triggers both
PSR1.Classes.ClassDeclaration.MultipleClasses and a misfire of
Magento2.NamingConvention.ReservedWords.ForbiddenAsNameSpace (the
sniff misreads `false`/`null` inside bracketed namespaces as namespace
fragments).

CI never caught this because the phpcs job ignored Test/.

Changes:
- Test/Unit/_bootstrap.php: PHPUnit bootstrap that loads composer's
  autoloader (robust to both repo-local and Magento-vendor install
  layouts) and then recursively requires every file under
  Test/Unit/_stubs/. phpunit.xml now points at this bootstrap.
- Test/Unit/_stubs/: 18 one-class-per-file Magento stubs in a
  directory tree mirroring their namespace, each guarded by
  class_exists / interface_exists / function_exists with
  $autoload=false so they no-op when the real Magento autoloader has
  already loaded the real class.
- Test/Unit/Observer/Stubs/: 6 test-helper stubs extracted from
  SaveOrderMarketingConsentTest (TestableSaveOrderMarketingConsent,
  StubOrder, StubAddress, StubQuote, StubEvent, StubObserver). PSR-4
  autoloaded under Klaviyo\Reclaim\Test\Unit\Observer\Stubs.
- Test/Unit/Setup/Patch/Data/Stubs/SelectStub.php: query-builder
  shim extracted from MigrateSmsToMobileConsentTest.
- 5 refactored test files: single-line `namespace ...;` syntax,
  one class per file, use imports for the extracted stubs.
- .github/workflows/continuous_integration.yml: drop the */Test/*
  exclusion from the phpcs ignore list so CI scans Test/ like the
  Marketplace command does.

Verification:
- All 27 tests in the 5 refactored files pass in the docker container.
- `phpcs --standard=Magento2 --severity=10` against the full plugin
  tree returns exit 0 (zero errors).
