+++ Test/Unit/_stubs/Klaviyo/Reclaim/Model/Config/Source/translation_function.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Klaviyo\Reclaim\Model\Config\Source;
+
+if (!function_exists('Klaviyo\Reclaim\Model\Config\Source\__')) {
+    function __($string)
+    {
+        return $string;
+    }
+}
