+++ Test/Unit/_stubs/Magento/Checkout/Api/Data/ShippingInformationInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Magento\Checkout\Api\Data;
+
+if (!interface_exists(\Magento\Checkout\Api\Data\ShippingInformationInterface::class, false)) {
+    interface ShippingInformationInterface
+    {
+        public function getExtensionAttributes();
+    }
+}
