+++ Test/Unit/_stubs/Magento/Checkout/Block/Checkout/LayoutProcessor.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Magento\Checkout\Block\Checkout;
+
+if (!class_exists(\Magento\Checkout\Block\Checkout\LayoutProcessor::class, false)) {
+    class LayoutProcessor
+    {
+    }
+}
