+++ Test/Unit/_stubs/Magento/Checkout/Model/ShippingInformationManagement.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Magento\Checkout\Model;
+
+if (!class_exists(\Magento\Checkout\Model\ShippingInformationManagement::class, false)) {
+    class ShippingInformationManagement
+    {
+    }
+}
