+++ Test/Unit/_stubs/Magento/Framework/DB/Adapter/AdapterInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Magento\Framework\DB\Adapter;
+
+if (!interface_exists(\Magento\Framework\DB\Adapter\AdapterInterface::class, false)) {
+    interface AdapterInterface
+    {
+        public function startSetup();
+        public function endSetup();
+        public function select();
+        public function fetchAll($select);
+        public function fetchOne($select);
+        public function insert($table, array $data);
+    }
+}
