+++ Test/Unit/_stubs/Magento/Framework/Event/Observer.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Magento\Framework\Event;
+
+if (!class_exists(\Magento\Framework\Event\Observer::class, false)) {
+    class Observer
+    {
+        public function getEvent()
+        {
+            return new Event();
+        }
+    }
+}
