+++ Test/Unit/_stubs/Magento/Framework/Event/ObserverInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Magento\Framework\Event;
+
+if (!interface_exists(\Magento\Framework\Event\ObserverInterface::class, false)) {
+    interface ObserverInterface
+    {
+        public function execute(Observer $observer);
+    }
+}
