+++ Test/Unit/_stubs/Magento/Framework/Option/ArrayInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Magento\Framework\Option;
+
+if (!interface_exists(\Magento\Framework\Option\ArrayInterface::class, false)) {
+    interface ArrayInterface
+    {
+        public function toOptionArray();
+    }
+}
