+++ Test/Unit/_stubs/Magento/Framework/Setup/Patch/DataPatchInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Magento\Framework\Setup\Patch;
+
+if (!interface_exists(\Magento\Framework\Setup\Patch\DataPatchInterface::class, false)) {
+    interface DataPatchInterface
+    {
+        public function apply();
+        public function getAliases();
+        public static function getDependencies();
+    }
+}
