+++ Test/Unit/_stubs/Magento/Framework/Setup/Patch/PatchVersionInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Magento\Framework\Setup\Patch;
+
+if (!interface_exists(\Magento\Framework\Setup\Patch\PatchVersionInterface::class, false)) {
+    interface PatchVersionInterface
+    {
+        public static function getVersion();
+    }
+}
