+++ Test/Unit/_stubs/Magento/Quote/Model/QuoteRepository.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Magento\Quote\Model;
+
+if (!class_exists(\Magento\Quote\Model\QuoteRepository::class, false)) {
+    class QuoteRepository
+    {
+        public function getActive($cartId)
+        {
+            return null;
+        }
+    }
+}
