+++ phpunit.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <phpunit
-    bootstrap="vendor/autoload.php"
+    bootstrap="Test/Unit/_bootstrap.php"
     colors="true"
     verbose="true"
 >
